Move LazyWait to own directory
