+++ src/index.ts
@@ -3,9 +3,11 @@ export { MODE, PRIORITY } from './constants';
 export { LooselyLazy as default } from './init';
 
 export { isLoaderError, lazyForPaint, lazyAfterPaint, lazy } from './lazy';
-
 export type { ClientLoader, Loader, ServerLoader } from './lazy';
 export type { LazyOptions, LazyComponent } from './lazy';
+
+export { LazyWait } from './lazy-wait';
+export type { LazyWaitProps } from './lazy-wait';
 
 export { getAssetUrlsFromId } from './manifest';
 export type { Asset, Manifest } from './manifest';
@@ -13,6 +15,5 @@ export type { Asset, Manifest } from './manifest';
 export { LazySuspense } from './suspense';
 export type { Fallback, LazySuspenseProps } from './suspense';
 
-export { LazyWait, useLazyPhase } from './phase';
-export type { LazyWaitProps } from './phase';
+export { useLazyPhase } from './phase';
 export type { Config } from './types';
