+++ src/lazy-wait/__tests__/lazy-wait.typescript.test.tsx
