+++ src/lazy-wait/index.tsx
@@ -1,0 +1,2 @@
+export { LazyWait } from './main';
+export type { LazyWaitProps } from './main';
