--- src/phase/types.ts
@@ -1,6 +1,0 @@
-import { ReactNode } from 'react';
-
-export type LazyWaitProps = {
-  until: boolean;
-  children: ReactNode;
-};
